fix bug #4 -  not updating table stock after update

+++ JustRipe/ViewModels/FertiliserViewModel.cs
@@ -82,8 +82,8 @@ namespace JustRipe.ViewModels
          set { _unit = value; OnPropertyChanged(nameof(Unit)); }
       }
 
-      private decimal _price;
-      public decimal Price
+      private double _price;
+      public double Price
       {
          get { return _price; }
          set { _price = value; OnPropertyChanged(nameof(Price)); }
